+++ src/style-spec/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 16.0.1
+
+### ✨ Features and improvements
+* Add `fill-composite-operation` fill paint property [#1191](https://github.com/maplibre/maplibre-gl-js/pull/1191)
+
 ## 16.0.0
 
 ### Breaking changes